requirements.txt
python-telegram-bot==13.15
edge-tts>=6.1.0
pydub>=0.25.1
ffmpeg-python>=0.2.1